extract properties popup component
